Move data model to its own directory
Dimensions, TestCase and its utility classes are the "Data Model" of
this app and belong together.
